refactory: create index file to export all components
