+++ src/components/index.jsx
@@ -1,0 +1,6 @@
+import Logo from "../components/Logo/";
+import WeatherInfo from "../components/WeatherInfo/";
+import WeatherDetails from "../components/WeatherDetails/";
+import WeatherIcon from "../components/WeatherIcon/";
+
+export { Logo, WeatherDetails, WeatherIcon, WeatherInfo }
